Remove magic +1 from QMenu::internalDelayedPopup
The position passed to QMenu::popup is always one pixel off causing
wrong menu placement under some circustances.

For example take this screens setup:
     1          2
[1920x1080][1920x1080]

If the menu is invoked in the right edge of the second screen the QPoint
is going to be 3840 which will be out of bound of the screen geomeetry.

Git history (qt.git inclusive) does not have any information on why this
magic +1 is in place.

Task-number: QTBUG-53813
Change-Id: Ic3b52ae2157b71b0438ebbfe3989f23660e302e9